+++ noxfile.py
@@ -30,7 +30,7 @@ nox.needs_version = ">=2025.10.16"
 nox.options.default_venv_backend = "uv"
 
 
-PYTHON_ALL_VERSIONS = ["3.10", "3.11", "3.12", "3.13", "3.14"]
+PYTHON_ALL_VERSIONS = ["3.10", "3.11", "3.12"]
 
 if os.environ.get("CI", None):
     nox.options.error_on_missing_interpreters = True
@@ -106,6 +106,7 @@ def minimums(session: nox.Session) -> None:
             session,
             install_args=["--resolution=lowest-direct"],
             run_args=["-Wdefault"],
+            extra_torch=True,
         )
         env = {"UV_PROJECT_ENVIRONMENT": session.virtualenv.location}
         session.run("uv", "tree", "--frozen", env=env)